+++ Core/Src/HAL_Drivers/HAL_SD.c
@@ -583,7 +583,7 @@ sd_status_t SD_Wave_Read(FIL *myFile, int32_t *IMUArray,  uint32_t WaveDirNo, ui
 }
 
 // Optimized SD_Wave_Read function
-sd_status_t SD_Wave_Read_Fast(FIL *myFile, int32_t *IMUArray, uint32_t WaveDirNo, uint32_t WaveLogNo, IMU_Data_SD_t inertialDataType, uint32_t *fpointer) 
+sd_status_t SD_Wave_Read_Fast(FIL *myFile, float *IMUArray, uint32_t WaveDirNo, uint32_t WaveLogNo, IMU_Data_SD_t inertialDataType, uint32_t *fpointer) 
 {
     // Buffer size adjusted to read larger blocks
     uint8_t dataBuf[4096];  // Adjust size based on available RAM
